requirements.txt
# Enhanced PPE Detection System Requirements
# Core ML and Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Data Processing and Utilities
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dataset and Augmentation
roboflow>=1.1.0
albumentations>=1.3.0

# API and HTTP
requests>=2.31.0
httpx>=0.25.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional GPU Support
# Uncomment if using CUDA
# torch-audio>=2.0.0
# torchaudio>=2.0.0